requirements.txt
attrs==17.4.0
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
enum34==1.1.6
Flask==0.12.2
funcsigs==1.0.2
futures==3.1.1
google-api-core==1.1.0
google-auth==1.4.1
google-cloud-core==0.28.1
google-cloud-logging==1.6.0
google-cloud-storage==1.8.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.10.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.1.0
pluggy==0.6.0
protobuf==3.5.2.post1
psycopg2-binary==2.7.4
py==1.5.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pytest==3.5.0
python-dotenv==0.8.2
pytz==2018.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
uWSGI==2.0.17
Werkzeug==0.14.1
